desktop_app/requirements.txt
pywebview==4.3.3
pyinstaller==5.13.2
